requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
python-multipart
timm